+++ bin/opub.dart
@@ -7,6 +7,6 @@
 import 'package:onepub/src/entry_point.dart';
 import 'package:onepub/src/my_runner.dart';
 
-void main(List<String> arguments) async {
+Future<void> main(List<String> arguments) async {
   await entrypoint(arguments, CommandSet.opub, 'onepub');
 }