+++ lib/src/api/version_check.dart
@@ -1,0 +1,6 @@
+import 'package:pub_semver/pub_semver.dart';
+
+class VersionCheck {
+  VersionCheck(this.serverVersion);
+  Version serverVersion;
+}