+++ lib/src/commands/login.dart
@@ -6,9 +6,8 @@ import 'dart:async';
 
 import 'package:args/command_runner.dart';
 import 'package:dcli/dcli.dart';
-import 'package:dcli/docker.dart';
 
-import '../exceptions.dart';
+import '../api/api.dart';
 import '../onepub_settings.dart';
 import '../util/bread_butter_auth.dart';
 import '../util/one_pub_token_store.dart';
@@ -45,52 +44,22 @@ class OnePubLoginCommand extends Command<int> {
     /// remote session will always be over ssh
     /// but local ones won't
 
-    checkForSSH();
-
-    checkForDocker();
-
     try {
+      await API().checkVersion();
+
       final bb = BreadButter();
-      final tempAuthTokenResponse = await bb.auth();
-      if (tempAuthTokenResponse == null) {
-        throw ExitException(
-            exitCode: 1, message: 'Invalid response. onePubToken not returned');
-      }
+      final auth = await bb.auth();
 
-      if (tempAuthTokenResponse.success) {
-        final onepubToken =
-            tempAuthTokenResponse.data['onePubToken'] as String?;
-        final firstLogin = tempAuthTokenResponse.data['firstLogin'] as bool?;
-        final operatorEmail =
-            tempAuthTokenResponse.data['operatorEmail'] as String?;
-        final organisationName =
-            tempAuthTokenResponse.data['organisationName'] as String?;
-        final obfuscatedOrganisationId =
-            tempAuthTokenResponse.data['obfuscatedOrganisationId'] as String?;
-        if (onepubToken == null ||
-            firstLogin == null ||
-            organisationName == null ||
-            operatorEmail == null ||
-            obfuscatedOrganisationId == null) {
-          print(tempAuthTokenResponse.data);
-          throw ExitException(
-              exitCode: 1,
-              message: 'Invalid response. missing authorization data');
-        }
+      OnePubTokenStore().save(
+          onepubToken: auth.onepubToken,
+          organisationName: auth.organisationName,
+          obfuscatedOrganisationId: auth.obfuscatedOrganisationId,
+          operatorEmail: auth.operatorEmail);
 
-        OnePubTokenStore().save(
-            onepubToken: onepubToken,
-            organisationName: organisationName,
-            obfuscatedOrganisationId: obfuscatedOrganisationId,
-            operatorEmail: operatorEmail);
-
-        showWelcome(
-            firstLogin: firstLogin,
-            organisationName: organisationName,
-            operator: operatorEmail);
-      } else {
-        showError(tempAuthTokenResponse);
-      }
+      showWelcome(
+          firstLogin: auth.firstLogin,
+          organisationName: auth.organisationName,
+          operator: auth.operatorEmail);
     } on FetchException {
       printerr(red('Unable to connect to '
           '${OnePubSettings.use.onepubApiUrl}. '
@@ -99,41 +68,41 @@ class OnePubLoginCommand extends Command<int> {
     return 0;
   }
 
-  void checkForSSH() {
-    if (inSSH()) {
-      throw ExitException(exitCode: -1, message: """
-${red('onepub login will not work from an ssh shell.')}
+//   void checkForSSH() {
+//     if (inSSH()) {
+//       throw ExitException(exitCode: -1, message: """
+// ${red('onepub login will not work from an ssh shell.')}
 
-Instead:
-Exit your ssh session and run:
-${green('onepub export')}
+// Instead:
+// Exit your ssh session and run:
+// ${green('onepub export')}
 
-Restart your ssh session and run:
-${green('onepub import --ask')}
+// Restart your ssh session and run:
+// ${green('onepub import --ask')}
 
-See the documentation for full details and alternate techniques:
-${orange('https://docs.onepub.dev/guides/ssh')}
-""");
-    }
-  }
+// See the documentation for full details and alternate techniques:
+// ${orange('https://docs.onepub.dev/guides/ssh')}
+// """);
+//     }
+//   }
 
-  void checkForDocker() {
-    if (DockerShell.inDocker) {
-      throw ExitException(exitCode: -1, message: """
-${red('onepub login will not work within a Docker shell.')}
-    
-Instead:
-Exit your docker session and run:
-${green('onepub export')}
+//   void checkForDocker() {
+//     if (DockerShell.inDocker) {
+//       throw ExitException(exitCode: -1, message: """
+// ${red('onepub login will not work within a Docker shell.')}
 
-Restart your docker session and run:
-${green('onepub import --ask')}
+// Instead:
+// Exit your docker session and run:
+// ${green('onepub export')}
 
-See the documentation for full details and alternate techniques:
-${orange('https://docs.onepub.dev/guides/docker')}
-""");
-    }
-  }
+// Restart your docker session and run:
+// ${green('onepub import --ask')}
+
+// See the documentation for full details and alternate techniques:
+// ${orange('https://docs.onepub.dev/guides/docker')}
+// """);
+//     }
+//   }
 
   void showError(EndpointResponse endPointResponse) {
     final error = endPointResponse.data['message']! as String;