+++ lib/src/commands/logout.dart
@@ -33,11 +33,12 @@ The logout command takes no arguments. Found ${argResults!.rest.join(',')}.
     }
 
     if (OnePubTokenStore().isLoggedIn) {
+      await API().checkVersion();
       final response = await API().logout();
       OnePubTokenStore().clearOldTokens();
 
       if (!response.success) {
-        throw ExitException(exitCode: 1, message: response.errorMessage!);
+        throw ExitException(exitCode: 1, message: response.errorMessage);
       }
 
       print(green('You have been logged out of the OnePub CLI for '