+++ lib/src/util/send_command.dart
@@ -61,7 +61,7 @@ You must be logged in to run this command.
 run: onepub login
   ''');
     }
-    final onepubToken = OnePubTokenStore().fetch();
+    final onepubToken = OnePubTokenStore().load();
 
     _headers.addAll({'authorization': onepubToken});
   }
@@ -198,8 +198,8 @@ class EndpointResponse {
   void _parseCli() {
     final decodedResponse = _bodyAsJsonMap(_body);
 
-    if (decodedResponse.keys.contains('success')) {
-      _data = decodedResponse['success'] as Map<String, Object?>;
+    if (decodedResponse.keys.contains('body')) {
+      _data = decodedResponse['body'] as Map<String, Object?>;
       _success = true;
     } else if (decodedResponse.keys.contains('error')) {
       _data = decodedResponse['error'] as Map<String, Object?>;