+++ lib/src/util/token_export_file.dart
@@ -32,6 +32,7 @@ class TokenExportFile {
   late final String pathToExportFile;
   late final SettingsYaml settings;
 
+  // ignore: discarded_futures
   void save() => waitForEx(settings.save());
 
   bool get hasToken => settings.validString(onepubTokenKey);